Refactor valueType from "Integer" to "Numeric" across multiple converters and scenarios
- Updated valueType in ConverterStandard, Converter_BangleJS2BLE, Converter_BulpSensorBLE, Converter_BulpWebRobo321, Converter_BulpLoRaRobo666, and various Zigbee converters to use "Numeric" instead of "Integer".
- Adjusted ScenarioEngine to reflect changes in valueType handling.
- Modified API documentation to replace "Integer" with "Numeric" in valueType enumerations.
- Ensured consistency in valueType usage across all relevant files.

+++ bridge - http/converters/Converter_BulpWebRobo321.js
@@ -17,7 +17,7 @@ class Converter_BulpWebRobo321 extends ConverterStandard {
             name:       "voltage",
             read:       true,
             anyValue:   0,
-            valueType:  "Integer"
+            valueType:  "Numeric"
         };
 
         this.properties[1] = {
